+++ src/lib/kotlin/slatekit-actors/src/main/kotlin/slatekit/actors/Workable.kt
@@ -1,0 +1,12 @@
+package slatekit.actors
+
+import kotlinx.coroutines.Job
+
+interface Workable {
+    /**
+     * Launches work ( e.g. by processing all items in a channel )
+     */
+    suspend fun work(): Job
+}
+
+
